+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog
 
 - New function ``utils.grid2grid`` to move between different model grids;
   factor-out ``_interp3d`` from ``utils.get_receiver`` for that.
+- New jitted function ``njitted.volume_average`` for interpolation using the
+  volume-average technique. Default method in ``utils.grid2grid``.
 
 
 *v0.7.0* : H-field